requirements/common.txt
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
django-cors-headers==4.0.0
paytm-pg==1.1.1
razorpay==1.3.0
qrcode==7.4.2
imgkit==1.2.3
django-import-export==3.2.0
cryptography==41.0.2
celery==5.3.1
redis==4.6.0
hiredis==2.2.3